Log a message to note that when a receiver is waiting for packets
Before the receiver object in the frontend can be shut down, the Windows
network stack must return all buffers passed up. The code therefore
waits for this to happen during the VIF Disable method. Add logging to
note that the wait is occuring as it can take a significant amount of time.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>