Use prefix matching for IPv6 address copy code
...and also fix registry path and add support for similar IPv4 address
copying.

Windows puts static IP address values in some oddly named and located
registry keys. The key names are just GUIDs and the only useful part of
the values names (that allows use to find them) is that they prefix
match the interface LUID rendered as a 16 digit hex string.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>