Avoid holding a reference to the UNLPLUG interface
If the system goes into S3 or S4 then currently the reference will still be
held. This is not problematic for this interface but it is not desirable to
hold references in this circumstance and for other interfaces it can lead
to bugchecks.

This patch adds a function, PdoRequestUnplug(), which only acquires the
interface as needed and then makes use of that.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>