Add frontend code for the new netif control ring
My recent patches to Xen's netif.h specify a control ring that can be
used by a frontend driver to configure packet hashing and steering in
a backend.
This patch adds the necessary code to XENVIF to drive this new ring,
however the rest of the code to link this up to new VIF interface
functionality (so that it may be used by XENNET) is deferred to a
subsequent patch.

This patch also pulls in an updated EVTCHN interface from XENBUS and
corrects the PDO binding.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>