Send transmit side hash value to the backend
Backends capable of supporting packet hashing on the guest receive side
are also capable of accepting hash values in an extra info fragment on
the guest transmit side. This patch adds code to construct and send the
extra info fragment if the packet metadata contains hash information and
the backend is able to process it.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>