Remove erroneous code to re-acquire backend path on resume
Commit 765b7a6a added code to release and re-acquire the backend
path (and domid) on resume from suspend. That code falsely
assumes that the frontend is not in a closed state at the time.
It's also not clear why it was added in the first place so
this patch removes it again, as it causes a BSOD when the
frontend is in the closed state.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>