Prevent VIF from being incorrectly considered offline...
...one resume from suspend.

My previous reversion of 765b7a6a exposed the original reason the
reverted patch.

On resume from suspend, the backend will have been re-created by the
toolstack and hence the frontend detects it as being offline. This
causes the frontend to request ejection of the disk (thinking that
the reason for the backend going offline is because the VIF is being
unplugged).

This patch works around this problem by introducing an early suspend
callback to note that the frontend has actually gone offline, which
will prevent the ejection from being requested.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>