Make sure transmitter DPC does not try to unmask non-existent evtchn
In the case where the backend does not support split event channels (one
for RX and one for TX) then the transmitter code in XENVIF does not open
the combined channel; it is handled by the receiver code. So, the
transmitter DPC should not make any attempt to unmask the event channel at
the end of processing.

This patch adds the necessary check for FrontendIsSplit before doing the
unmask and also removes an unnecessary check for FrontendIsSplit in the
transmitter ring enable function.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>
Reported-by: Eytan Heidingsfeld <eytanh@gmail.com>
Tested-by: Eytan Heidingsfeld <eytanh@gmail.com>