Remove settings code
Unfortunately, after extensive testing of different scenarios, it seems to
be impossible to properly steal and restore network stack bindings from
within the XENVIF driver. Windows has already sampled some of the values
we need to modify and thus we cannot successfully modify them.

It appears the only way we can successfully use settings from an emulated
device is to do what the settings code was attempting to do, but do it
in XENNET's co-installer during the pre-install and pre-removal phases.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>