Revert all settings stealing patches
Windows has completely defeated me. There appears to be no way to
duplicate the stack binding of the emulated device without that binding
being removed if XENNET is un-installed, despite clearing all the
relevant registry keys in the DIF_REMOVE pre-process phase.

So, this patch reverts the code back to using the settings copy
mechanism employed in the 8.1 source.

The patches reverted (with some fix-up) are:

32bd2878 "Remove settings code"
70f4bb96 "Re-work identity stealing code"
04c391d9 "Replace copying network settings with identity stealing"

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>