Don't restore settings from emulated device more than once
If, for some reason, the VM boots with emulated networking then network
settings will be saved from the emulated network device. Then, when the
VM reboots with PV networking, those settings are restored to the PV
network device regardless of whether it was freshly installed or has
been around for some time.

This patch makes sure that settings are restored only to a freshly
installed PV device.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>