Incorporate revision ids from staging-8.1
Since it should be possible to upgrade from an 8.1 version of XENVIF
to an 8.2 version, at least the release revision of the 8.1 version
(0x0800000A) should be present in the 8.2 revision table, otherwise
driver upgrade will be vetoed by the co-installer if an 8.1 XENNET
is bound.

This patch incorporates all missing 8.1 revision ids.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>