Partially revert commit ab655bb1 "Make sure XENBUS interfaces...
...are released when going into S4".

Not all interfaces need to be released (since they don't all depend on
Xen) and crucially the receiver packet caches CANNOT be destroyed (and
hence the CACHE interface CANNOT be released) because there may be packets
outstanding in the stack... not necessarily in an S4 transtion, but
across a suspend/resume (which involves the same frontend state
transitions).

This patch also increases the log level of a couple of messages
emitted during frontend state transtions from 'trace' to 'info'.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>