Limit retry in ControllerGetResponse
During unplug, NDIS may issue a hash reconfiguration. However, the
backend may disappear before the request has had a chance to complete.
This is observable by e.g. adding a delay at the beginning of
VifReceiverSetHashAlgorithm and unplugging. If this happens, Xenvif will
get stuck waiting forever in ControllerGetResponse, at DISPATCH_LEVEL to
boot.

Limit the EvtchnWait retry to 10 attempts (10ms each to keep the
previous 100ms limit) and return an error code to the caller if the wait
has failed.

Signed-off-by: Tu Dinh <ngoc-tu.dinh@vates.tech>